cleanup of files, example graphics

+++ Render.php
@@ -1,17 +1,10 @@
 <?php
 class Render {
-	private $LATEX_PATH = "/usr/bin/latex";
-	private $DVIPS_PATH = "/usr/bin/dvips";
-	private $CONVERT_PATH = "/usr/bin/convert";
-	private $TMP_DIR ="./tmp";
-
-	private $CACHE_DIR = "";
-	private $URL_PATH = "";
+	private $asy_images = '';
 	private $problem_section;
-	private $img_width;
 
-	public function __construct($img_width = 300) {
-		$this->img_width = $img_width;
+	public function __construct($asy_images = false) {
+		$this->asy_images = $asy_images;
 	}// end constructor
 
 	/*
@@ -48,8 +41,6 @@ return <<<EOS
 \\pagestyle{empty}
 \\thispagestyle{empty}
 
-\\definecolor{hebg}{RGB}{246,255,255}
-
 \\begin{document}
 	$thunk
 \\end{document}
@@ -60,37 +51,30 @@ EOS;
 	* Purpose: transform raw code from file to TeX-renderable code
 	*
 	* @param string text: original problem containing TeX code
-	* @param string path: destination of rendered TeX image
+	* @param integer width: final width of rendered TeX image in pixels
 	*
 	* @return string: text containing TeX code replaced by image tags
 	*/
-	public function transform($text, $path) {
-		require_once('Asymptote.php');
+	public function transform($text, $width) {
+		$asyHash = '';
 
-		// replace \begin{tex} and \end{tex} tags with $
-		#$text = preg_replace("/\\\\begin\\{tex\\}/", "$", $text);
-		#$text = preg_replace("/\\\\end\\{tex\\}/", "$", $text);
-
-		// replace [] tags with $$
-		#$text = preg_replace("/\\\\\[(.*?)\\\\\]/", "$$ $1 $$", $text);
+		if ($this->asy_images)
+			require_once('Asymptote.php');
 
 		// change \$ to #dol temporarily
 		$text = preg_replace("/\\\\\\$/", "#dol", $text);
 
-		// change <br/> tags to \\ (TeX newline)
-		$text = preg_replace("/(\s+)?<br(\s*\/?)?>(\s+)?/", " \\\\\\\\ ", $text);
-
-		// render asymptote to EPS
-		$asy = new Asymptote();
-		list($text, $asyHash) = $asy->transform($text, $path);
-		unset($asy);
-
-		$this->CACHE_DIR = $path;
+		if ($this->asy_images) {
+			// render asymptote to EPS
+			$asy = new Asymptote();
+			list($text, $asyHash) = $asy->transform($text);
+			unset($asy);
+		}// end if
 
 		// create cache directory if DNE, then process latex
-		if (is_dir($this->CACHE_DIR) ||
-		mkdir($this->CACHE_DIR, 0755, true)) {
-			$thunk = "\\begin{varwidth}{{$this->img_width}px}$text\\end{varwidth}";
+		if (is_dir(Config::CACHE_DIR) ||
+		mkdir(Config::CACHE_DIR, 0755, true)) {
+			$thunk = "\\begin{varwidth}{{$width}px}$text\\end{varwidth}";
 			$thunk = "$\\textrm{" . $thunk . "}$";
 
 			// replace #dol with \$ again
@@ -98,7 +82,7 @@ EOS;
 
 			// append time to thunk before md5 for better randomization
 			$hash = md5($thunk . time());
-			$full_name = $this->CACHE_DIR . "/$hash.png";
+			$full_name = Config::CACHE_DIR . "/$hash.png";
 
 			// do not overwrite file if exists
 			if (!is_file($full_name)) {
@@ -121,21 +105,28 @@ EOS;
 	private function render_latex($thunk, $hash) {
 		$thunk = $this->wrap($thunk);
 		$current_dir = getcwd();
-		chdir($this->TMP_DIR);
+		chdir(Config::TMP_DIR);
 
 		// create temporary LaTeX file
-		#fp = fopen($this->TMP_DIR . "/$hash.tex", "w+");
+		#$fp = fopen(Config::TMP_DIR . "/$hash.tex", "w+");
 		$fp = fopen("./$hash.tex", "w+");
 		fputs($fp, $thunk);
 		fclose($fp);
 
 		// convert from latex -> dvi -> ps -> png
-		$command = $this->LATEX_PATH . " --shell-escape --interaction=nonstopmode $hash.tex; " . $this->DVIPS_PATH . " -E $hash -o $hash.ps; " . $this->CONVERT_PATH . " -density 850 $hash.ps -colorspace RGB -filter Lanczos -resize 20% -colorspace sRGB $hash.png";
+		if ($this->asy_images)
+			$command = Config::LATEX_PATH . " --shell-escape --interaction=nonstopmode $hash.tex; " . Config::DVIPS_PATH . " -E $hash -o $hash.ps; " . Config::CONVERT_PATH . " -density 850 $hash.ps -colorspace RGB -filter Lanczos -resize 20% -colorspace sRGB $hash.png";
+		
+		// convert from latex -> dvi -> png
+		else
+			$command = Config::LATEX_PATH . " --shell-escape $hash.tex; " . Config::DVIPNG_PATH . " -q -T tight -D 170 -bg Transparent $hash -o $hash.png";// > /dev/null 2>&1";
+
 		exec($command);
 
 		chdir($current_dir);
+
 		// copy the file to the cache directory
-		copy($this->TMP_DIR . "/$hash.png", $this->CACHE_DIR . "/$hash.png");
+		copy(Config::TMP_DIR . "/$hash.png", Config::CACHE_DIR . "/$hash.png");
 	}// end render_latex
 
 	/*
@@ -147,13 +138,15 @@ EOS;
 	*/
 	private function cleanup($hash, $asyHash) {
 		$current_dir = getcwd();
-		chdir($this->TMP_DIR);
+		chdir(Config::TMP_DIR);
 		file_exists("$hash.tex") && unlink("$hash.tex");
 		file_exists("$hash.aux") && unlink("$hash.aux");
 		file_exists("$hash.log") && unlink("$hash.log");
 		file_exists("$hash.dvi") && unlink("$hash.dvi");
-		file_exists("$hash.ps") && unlink("$hash.ps");
 		file_exists("$hash.png") && unlink("$hash.png");
+
+		if ($this->asy_images)
+			file_exists("$hash.ps") && unlink("$hash.ps");
 
 		if ($asyHash != '')
 			file_exists("$asyHash.eps") && unlink("$asyHash.eps");
